[Refactor] cell configuration 코드 분리
cellRegistrationHandler안에 있던  cell configuration 코드를 분리하는 작업
